requirements.txt
PyQt5>=5.15
requests>=2.31
PyYAML>=6.0
Pillow>=10.0
numpy>=1.26
torch>=2.2
torchvision>=0.17
git+https://github.com/facebookresearch/segment-anything.git
